release 69

+++ app/src/main/java/com/yamilab/animalsounds/AnimalAdapter.java
@@ -1,23 +1,15 @@
 package com.yamilab.animalsounds;
 
 import android.content.Context;
-import android.content.Intent;
-import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Build;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.MemoryCategory;
 import com.bumptech.glide.Priority;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.bumptech.glide.request.target.Target;
@@ -27,7 +19,6 @@ import com.google.firebase.storage.StorageReference;
 import java.util.ArrayList;
 
 import static com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions.withCrossFade;
-import static java.security.AccessController.getContext;
 
 /**
  * Created by Misha on 25.02.2018.
@@ -222,7 +213,7 @@ public class AnimalAdapter extends RecyclerView.Adapter<AnimalAdapter.ViewHolder
 
                          .skipMemoryCache(true)
                         .diskCacheStrategy(DiskCacheStrategy.ALL)
-                        .override((int) screenWidth, Target.SIZE_ORIGINAL)
+                        .override(screenWidth, Target.SIZE_ORIGINAL)
                         .fitCenter()
                         // .thumbnail()
                         //.error(R.mipmap.ic_launcher)
